requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
altair==5.5.0
    # via streamlit
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.16.1
    # via torch
fonttools==4.55.3
    # via matplotlib
fsspec==2024.10.0
    # via torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
    #   torch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.3
    # via
    #   seaborn
    #   ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==1.17.0
    # via altair
networkx==3.4.2
    # via torch
numpy==2.2.0
    # via
    #   car-detection-web (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pydeck
    #   scipy
    #   seaborn
    #   streamlit
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.10.0.84
    # via ultralytics
packaging==24.2
    # via
    #   altair
    #   matplotlib
    #   streamlit
pandas==2.2.3
    # via
    #   car-detection-web (pyproject.toml)
    #   seaborn
    #   streamlit
    #   ultralytics
pillow==11.0.0
    # via
    #   matplotlib
    #   streamlit
    #   torchvision
    #   ultralytics
protobuf==5.29.1
    # via streamlit
psutil==6.1.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyarrow==18.1.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via ultralytics
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   streamlit
    #   ultralytics
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scipy==1.14.1
    # via ultralytics
seaborn==0.13.2
    # via ultralytics
six==1.17.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.41.0
    # via car-detection-web (pyproject.toml)
sympy==1.13.1
    # via torch
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
torch==2.5.1
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.20.1
    # via ultralytics
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via ultralytics
typing-extensions==4.12.2
    # via
    #   altair
    #   streamlit
    #   torch
tzdata==2024.2
    # via pandas
ultralytics==8.3.49
    # via car-detection-web (pyproject.toml)
ultralytics-thop==2.0.13
    # via ultralytics
urllib3==2.2.3
    # via requests
watchdog==6.0.0
    # via streamlit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
